Replace MapDefinitions
